+++ packages/core/src/js/tracing/reactnavigation.ts
@@ -2,7 +2,6 @@
 import type { Client, Integration, Span } from '@sentry/core';
 import {
   addBreadcrumb,
-  getActiveSpan,
   getClient,
   isPlainObject,
   logger,
@@ -15,24 +14,22 @@ import {
 } from '@sentry/core';
 
 import { getAppRegistryIntegration } from '../integrations/appRegistry';
-import type { NewFrameEvent } from '../utils/sentryeventemitter';
-import type { SentryEventEmitterFallback } from '../utils/sentryeventemitterfallback';
-import { createSentryFallbackEventEmitter } from '../utils/sentryeventemitterfallback';
 import { isSentrySpan } from '../utils/span';
 import { RN_GLOBAL_OBJ } from '../utils/worldwide';
+import type { UnsafeAction } from '../vendor/react-navigation/types';
 import { NATIVE } from '../wrapper';
 import { ignoreEmptyBackNavigation } from './onSpanEndUtils';
 import { SPAN_ORIGIN_AUTO_NAVIGATION_REACT_NAVIGATION } from './origin';
 import type { ReactNativeTracingIntegration } from './reactnativetracing';
 import { getReactNativeTracingIntegration } from './reactnativetracing';
-import { SEMANTIC_ATTRIBUTE_SENTRY_SOURCE } from './semanticAttributes';
+import { SEMANTIC_ATTRIBUTE_NAVIGATION_ACTION_TYPE, SEMANTIC_ATTRIBUTE_SENTRY_SOURCE } from './semanticAttributes';
 import {
   DEFAULT_NAVIGATION_SPAN_NAME,
   defaultIdleOptions,
   getDefaultIdleNavigationSpanOptions,
   startIdleNavigationSpan as startGenericIdleNavigationSpan,
 } from './span';
-import { manualInitialDisplaySpans, startTimeToInitialDisplaySpan, updateInitialDisplaySpan } from './timetodisplay';
+import { addTimeToInitialDisplayFallback } from './timeToDisplayFallback';
 export const INTEGRATION_NAME = 'ReactNavigation';
 
 const NAVIGATION_HISTORY_MAX_SIZE = 200;
@@ -69,6 +66,13 @@ interface ReactNavigationIntegrationOptions {
    * @default false
    */
   enableTimeToInitialDisplayForPreloadedRoutes: boolean;
+
+  /**
+   * Whether to use the dispatched action data to populate the transaction metadata.
+   *
+   * @default false
+   */
+  useDispatchedActionData: boolean;
 }
 
 /**
@@ -84,15 +88,16 @@ export const reactNavigationIntegration = ({
   enableTimeToInitialDisplay = false,
   ignoreEmptyBackNavigationTransactions = true,
   enableTimeToInitialDisplayForPreloadedRoutes = false,
+  useDispatchedActionData = false,
 }: Partial<ReactNavigationIntegrationOptions> = {}): Integration & {
   /**
    * Pass the ref to the navigation container to register it to the instrumentation
    * @param navigationContainerRef Ref to a `NavigationContainer`
    */
   registerNavigationContainer: (navigationContainerRef: unknown) => void;
+  options: ReactNavigationIntegrationOptions;
 } => {
   let navigationContainer: NavigationContainer | undefined;
-  let newScreenFrameEventEmitter: SentryEventEmitterFallback | undefined;
 
   let tracing: ReactNativeTracingIntegration | undefined;
   let idleSpanOptions: Parameters<typeof startGenericIdleNavigationSpan>[1] = defaultIdleOptions;
@@ -106,8 +111,6 @@ export const reactNavigationIntegration = ({
   let recentRouteKeys: string[] = [];
 
   if (enableTimeToInitialDisplay) {
-    newScreenFrameEventEmitter = createSentryFallbackEventEmitter();
-    newScreenFrameEventEmitter.initAsync();
     NATIVE.initNativeReactNavigationNewFrameTracking().catch((reason: unknown) => {
       logger.error(`${INTEGRATION_NAME} Failed to initialize native new frame tracking: ${reason}`);
     });
@@ -207,7 +210,30 @@ export const reactNavigationIntegration = ({
    * It does not name the transaction or populate it with route information. Instead, it waits for the state to fully change
    * and gets the route information from there, @see updateLatestNavigationSpanWithCurrentRoute
    */
-  const startIdleNavigationSpan = (): void => {
+  const startIdleNavigationSpan = (unknownEvent?: unknown): void => {
+    const event = unknownEvent as UnsafeAction | undefined;
+    if (useDispatchedActionData && event?.data.noop) {
+      logger.debug(`${INTEGRATION_NAME} Navigation action is a noop, not starting navigation span.`);
+      return;
+    }
+
+    const navigationActionType = useDispatchedActionData ? event?.data.action.type : undefined;
+    if (
+      useDispatchedActionData &&
+      [
+        // Process common actions
+        'PRELOAD',
+        'SET_PARAMS',
+        // Drawer actions
+        'OPEN_DRAWER',
+        'CLOSE_DRAWER',
+        'TOGGLE_DRAWER',
+      ].includes(navigationActionType)
+    ) {
+      logger.debug(`${INTEGRATION_NAME} Navigation action is ${navigationActionType}, not starting navigation span.`);
+      return;
+    }
+
     if (latestNavigationSpan) {
       logger.log(`${INTEGRATION_NAME} A transaction was detected that turned out to be a noop, discarding.`);
       _discardLatestTransaction();
@@ -221,11 +247,13 @@ export const reactNavigationIntegration = ({
       idleSpanOptions,
     );
     latestNavigationSpan?.setAttribute(SEMANTIC_ATTRIBUTE_SENTRY_ORIGIN, SPAN_ORIGIN_AUTO_NAVIGATION_REACT_NAVIGATION);
+    latestNavigationSpan?.setAttribute(SEMANTIC_ATTRIBUTE_NAVIGATION_ACTION_TYPE, navigationActionType);
     if (ignoreEmptyBackNavigationTransactions) {
       ignoreEmptyBackNavigation(getClient(), latestNavigationSpan);
     }
 
     if (enableTimeToInitialDisplay) {
+      NATIVE.setActiveSpanId(latestNavigationSpan?.spanContext().spanId);
       navigationProcessingSpan = startInactiveSpan({
         op: 'navigation.processing',
         name: 'Navigation dispatch to navigation cancelled or screen mounted',
@@ -265,6 +293,8 @@ export const reactNavigationIntegration = ({
       return undefined;
     }
 
+    addTimeToInitialDisplayFallback(latestNavigationSpan.spanContext().spanId, NATIVE.getNewScreenTimeToDisplay());
+
     if (previousRoute && previousRoute.key === route.key) {
       logger.debug(`[${INTEGRATION_NAME}] Navigation state changed, but route is the same as previous.`);
       pushRecentRouteKey(route.key);
@@ -276,32 +306,6 @@ export const reactNavigationIntegration = ({
     }
 
     const routeHasBeenSeen = recentRouteKeys.includes(route.key);
-    const startTtidForNewRoute = enableTimeToInitialDisplay && !routeHasBeenSeen;
-    const startTtidForAllRoutes = enableTimeToInitialDisplay && enableTimeToInitialDisplayForPreloadedRoutes;
-
-    let latestTtidSpan: Span | undefined = undefined;
-    if (startTtidForNewRoute || startTtidForAllRoutes) {
-      latestTtidSpan = startTimeToInitialDisplaySpan({
-        name: `${route.name} initial display`,
-        isAutoInstrumented: true,
-      });
-    }
-
-    const navigationSpanWithTtid = latestNavigationSpan;
-    if (latestTtidSpan) {
-      newScreenFrameEventEmitter?.onceNewFrame(({ newFrameTimestampInSeconds }: NewFrameEvent) => {
-        const activeSpan = getActiveSpan();
-        if (activeSpan && manualInitialDisplaySpans.has(activeSpan)) {
-          logger.warn('[ReactNavigationInstrumentation] Detected manual instrumentation for the current active span.');
-          return;
-        }
-
-        updateInitialDisplaySpan(newFrameTimestampInSeconds, {
-          activeSpan: navigationSpanWithTtid,
-          span: latestTtidSpan,
-        });
-      });
-    }
 
     navigationProcessingSpan?.updateName(`Navigation dispatch to screen ${route.name} mounted`);
     navigationProcessingSpan?.setStatus({ code: SPAN_STATUS_OK });
@@ -381,6 +385,13 @@ export const reactNavigationIntegration = ({
     name: INTEGRATION_NAME,
     afterAllSetup,
     registerNavigationContainer,
+    options: {
+      routeChangeTimeoutMs,
+      enableTimeToInitialDisplay,
+      ignoreEmptyBackNavigationTransactions,
+      enableTimeToInitialDisplayForPreloadedRoutes,
+      useDispatchedActionData,
+    },
   };
 };
 
@@ -392,6 +403,15 @@ export interface NavigationRoute {
 }
 
 interface NavigationContainer {
-  addListener: (type: string, listener: () => void) => void;
+  addListener: (type: string, listener: (event?: unknown) => void) => void;
   getCurrentRoute: () => NavigationRoute;
+}
+
+/**
+ * Returns React Navigation integration of the given client.
+ */
+export function getReactNavigationIntegration(
+  client: Client,
+): ReturnType<typeof reactNavigationIntegration> | undefined {
+  return client.getIntegrationByName<ReturnType<typeof reactNavigationIntegration>>(INTEGRATION_NAME);
 }